Adds Support For Detecting Half Rating Clicks
Updates the click handler to be able to detect when a half rating
should be assigned instead of a whole rating. Adds a halfRatings
property to turn on and off half step rating support.

Adds a check to remove the rating if the user attempts to assign the
exact same rating to the widget again.